requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.12
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
Babel==2.10.3
base58==2.1.1
beautifulsoup4==4.11.1
bitarray==2.6.0
bleach==5.0.1
cached-property==1.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
cytoolz==0.12.0
defusedxml==0.7.1
Django==4.1
django-ckeditor==6.5.0
django-cors-headers==3.13.0
django-email-verification==0.2.2
django-js-asset==2.0.0
django-phonenumber-field==6.3.0
django-tinymce==3.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
docutils==0.19
entrypoints==0.4
eth-abi==2.2.0
eth-account==0.5.9
eth-bloom==1.0.4
eth-hash==0.3.3
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-tester==0.6.0b6
eth-typing==2.3.0
eth-utils==1.9.5
fastjsonschema==2.16.1
frozenlist==1.3.1
futureproof==0.3.1
gunicorn==20.1.0
hexbytes==0.2.3
idna==3.3
imagesize==1.4.1
importlib-metadata==4.12.0
install==1.3.5
ipfshttpclient==0.8.0a2
Jinja2==3.1.2
jsonschema==4.14.0
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
lru-dict==1.1.8
lxml==4.9.1
MarkupSafe==2.1.1
mistune==0.8.4
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
nbclient==0.6.6
nbconvert==6.5.3
nbformat==5.4.0
nbsphinx==0.8.9
nest-asyncio==1.5.5
netaddr==0.8.0
packaging==21.3
pandocfilters==1.5.0
parsimonious==0.8.1
phonenumbers==8.12.54
protobuf==3.20.1
psutil==5.9.1
psycopg2-binary==2.9.3
py-ecc==5.2.0
py-evm==0.5.0a3
py-geth==3.8.0
pycryptodome==3.15.0
pyethash==0.1.27
Pygments==2.13.0
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
pysha3==1.0.2
python-dateutil==2.8.2
pytz==2022.2.1
pyzmq==23.2.1
requests==2.28.1
rlp==2.0.1
semantic-version==2.10.0
six==1.16.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
Sphinx==5.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
tinycss2==1.1.1
toolz==0.12.0
tornado==6.2
traitlets==5.3.0
trie==2.0.0a5
typing-extensions==3.10.0.2
ujson==5.4.0
urllib3==1.26.11
varint==1.0.2
web3==5.30.0
web3-ethereum-defi==0.10
webencodings==0.5.1
websockets==9.1
whitenoise==6.2.0
yarl==1.8.1
zipp==3.8.1
Pillow==9.0.0
boto3==1.24.70
django-storages==1.13.1
coinbase-commerce==1.0.1
python-decouple==3.6
stripe==5.0.0
ShopifyAPI==12.2.0